+++ source/extensions/omni.isaac.examples/omni/isaac/examples/humanoid/__init__.py
@@ -9,5 +9,6 @@
 
 # NOTE: Import here your extension examples to be propagated to ISAAC SIM Extensions startup
 
-from omni.isaac.examples.unitree_quadruped.quadruped_example import QuadrupedExample
-from omni.isaac.examples.unitree_quadruped.quadruped_example_extension import QuadrupedExampleExtension
+from omni.isaac.examples.humanoid.h1 import H1FlatTerrainPolicy
+from omni.isaac.examples.humanoid.humanoid_example import HumanoidExample
+from omni.isaac.examples.humanoid.humanoid_example_extension import HumanoidExampleExtension