+++ source/extensions/omni.isaac.quadruped/omni/isaac/quadruped/robots/spot.py
@@ -17,26 +17,26 @@ import omni.kit.commands
 import torch
 from omni.isaac.core.articulations import Articulation
 from omni.isaac.core.utils.prims import define_prim, get_prim_at_path
-from omni.isaac.core.utils.rotations import quat_to_rot_matrix
+from omni.isaac.core.utils.rotations import euler_to_rot_matrix, quat_to_euler_angles, quat_to_rot_matrix
 from omni.isaac.core.utils.stage import get_current_stage
+from omni.isaac.core.utils.types import ArticulationAction
 from omni.isaac.nucleus import get_assets_root_path
-from omni.isaac.quadruped.utils import LstmSeaNetwork
 from pxr import Gf
 
 
-class AnymalFlatTerrainPolicy:
-    """The ANYmal Robot running a Flat Terrain Locomotion Policy"""
+class SpotFlatTerrainPolicy:
+    """The Spot quadruped"""
 
     def __init__(
         self,
         prim_path: str,
-        name: str = "anymal",
+        name: str = "spot",
         usd_path: Optional[str] = None,
         position: Optional[np.ndarray] = None,
         orientation: Optional[np.ndarray] = None,
     ) -> None:
         """
-        Initialize anymal robot, import policy and actuator network.
+        Initialize robot and load RL policy.
 
         Args:
             prim_path {str} -- prim path of the robot on the stage
@@ -59,38 +59,32 @@ class AnymalFlatTerrainPolicy:
                 if assets_root_path is None:
                     carb.log_error("Could not find Isaac Sim assets folder")
 
-                asset_path = assets_root_path + "/Isaac/Robots/ANYbotics/anymal_c.usd"
+                asset_path = assets_root_path + "/Isaac/Robots/BostonDynamics/spot/spot.usd"
 
                 prim.GetReferences().AddReference(asset_path)
 
         self.robot = Articulation(prim_path=self._prim_path, name=name, position=position, orientation=orientation)
 
+        self._dof_control_modes: List[int] = list()
+
         # Policy
-        file_content = omni.client.read_file(assets_root_path + "/Isaac/Samples/Quadruped/Anymal_Policies/policy.pt")[2]
-
-        file = io.BytesIO(memoryview(file_content).tobytes())
-        self._policy = torch.jit.load(file)
-
-        # Policy Scales
-        self._base_vel_lin_scale = 1
-        self._base_vel_ang_scale = 1
-        self._action_scale = 0.5
-        self._default_joint_pos = np.array([0.0, 0.0, 0.0, 0.0, 0.4, -0.4, 0.4, -0.4, -0.8, 0.8, -0.8, 0.8])
-        self._previous_action = np.zeros(12)
-        self._policy_counter = 0
-
-        # Actuator network
         file_content = omni.client.read_file(
-            assets_root_path + "/Isaac/Samples/Quadruped/Anymal_Policies/sea_net_jit2.pt"
+            assets_root_path + "/Isaac/Samples/Quadruped/Spot_Policies/spot_policy.pt"
         )[2]
         file = io.BytesIO(memoryview(file_content).tobytes())
-        self._actuator_network = LstmSeaNetwork()
-        self._actuator_network.setup(file, self._default_joint_pos)
-        self._actuator_network.reset()
+
+        self._policy = torch.jit.load(file)
+        self._base_vel_lin_scale = 1
+        self._base_vel_ang_scale = 1
+        self._action_scale = 0.2
+        self._default_joint_pos = np.array([0.1, -0.1, 0.1, -0.1, 0.9, 0.9, 1.1, 1.1, -1.5, -1.5, -1.5, -1.5])
+        self._previous_action = np.zeros(12)
+        self._policy_counter = 0
+        self._decimation = 10
 
     def _compute_observation(self, command):
         """
-        Computes the the observation vector for the policy
+        Compute the observation vector for the policy
 
         Argument:
         command {np.ndarray} -- the robot command (v_x, v_y, w_z)
@@ -120,13 +114,11 @@ class AnymalFlatTerrainPolicy:
         obs[9] = self._base_vel_lin_scale * command[0]
         obs[10] = self._base_vel_lin_scale * command[1]
         obs[11] = self._base_vel_ang_scale * command[2]
-
         # Joint states
         current_joint_pos = self.robot.get_joint_positions()
         current_joint_vel = self.robot.get_joint_velocities()
         obs[12:24] = current_joint_pos - self._default_joint_pos
         obs[24:36] = current_joint_vel
-
         # Previous Action
         obs[36:48] = self._previous_action
 
@@ -141,35 +133,26 @@ class AnymalFlatTerrainPolicy:
         command {np.ndarray} -- the robot command (v_x, v_y, w_z)
 
         """
-        if self._policy_counter % 4 == 0:
+        if self._policy_counter % self._decimation == 0:
             obs = self._compute_observation(command)
             with torch.no_grad():
                 obs = torch.from_numpy(obs).view(1, -1).float()
                 self.action = self._policy(obs).detach().view(-1).numpy()
             self._previous_action = self.action.copy()
 
-        # The learning controller uses the order of
-        # FL_hip_joint FL_thigh_joint FL_calf_joint
-        # FR_hip_joint FR_thigh_joint FR_calf_joint
-        # RL_hip_joint RL_thigh_joint RL_calf_joint
-        # RR_hip_joint RR_thigh_joint RR_calf_joint
-        current_joint_pos = self.robot.get_joint_positions()
-        current_joint_vel = self.robot.get_joint_velocities()
+        action = ArticulationAction(joint_positions=self._default_joint_pos + (self.action * self._action_scale))
+        self.robot.apply_action(action)
 
-        joint_torques, _ = self._actuator_network.compute_torques(
-            current_joint_pos, current_joint_vel, self._action_scale * self.action
-        )
-
-        self.robot.set_joint_efforts(joint_torques)
         self._policy_counter += 1
 
     def initialize(self, physics_sim_view=None) -> None:
         """
-        Initialize the articulation interface, set up drive mode
+        Initialize robot the articulation interface, set up drive mode
         """
         self.robot.initialize(physics_sim_view=physics_sim_view)
         self.robot.get_articulation_controller().set_effort_modes("force")
-        self.robot.get_articulation_controller().switch_control_mode("effort")
+        self.robot.get_articulation_controller().switch_control_mode("position")
+        self.robot._articulation_view.set_gains(np.zeros(12) + 60, np.zeros(12) + 1.5)
 
     def post_reset(self) -> None:
         """